recalled history API with pagination

+++ app/src/main/java/uz/androbeck/virtualbank/domain/ui_models/history/InComeAndOutComeUIModel.kt
@@ -4,7 +4,7 @@ data class InComeAndOutComeUIModel(
     val type: String? = null,
     val from: String? = null,
     val to: String? = null,
-    val amount: Double = 0.0,
+    val amount: Float=0f,
     val time: Long = 0,
 
 )
